pyproject.toml
[project]
name = "ggg-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "flask>=3.1.2",
    "pillow>=11.3.0",
    "pillow-heif>=1.1.0",
    "pilmoji>=2.0.4",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uuid>=1.30",
    "werkzeug>=3.1.3",
]
